requirements.txt
pyaudio
pyautogui
opencv-python
pywinauto
python-vlc
matplotlib